chore: Filter trackers before query
Part of story #36833 search on semantics and always there fields even if not defined

The query join on tables `groups` and `tracker` only in order to filter
on which trackers we search. These 2 joins degrade query performance
when we search on many trackers.

This filter can be applied before directly on the list of tracker we use

*No functional changes expected*

Change-Id: Ic28c3ec30fc42797e1991ac31b0c34e3903ba305