refactor: Group submission date invalid checks together
Part of story #36833 search on semantics and always there fields even if
not defined

No functional change. CI should be happy.

Why ?
This finishes the refactoring of the structure of the
query validation to transition away from checks being grouped by
comparison type (equals, lesser than, etc).
We want to group checks by semantic / field type instead, as it is
easier to understand. Most of the time, we work "field by field", so it
makes more sense to have all the business rules for a given field in as
few files as possible, instead of having all the business rules for
"equals" comparisons being grouped together. Furthermore, usage of
inheritance makes it hard to keep a mental model of the validation rules
for a given field.

Change-Id: I32ec8b6d8e6db4521ee11d16ff1daee1b3063455