feat: Async events queue is now backed by the DB instead of Redis
This will allow us to rely on the async events everywhere as they will
not rely anymore on an optional piece of architecture and make the
deployments easier.

This is enabled by default but a temporary feature flag is available to
disable it in case of troubles.

Part of request #37566: Support the async event queue using the DB instead of Redis

Change-Id: If834d826d5e8d15cbe772c70c977fda11862f0d3