Introduce Artidoc Plugin
Base plugin to bootstrap the application.

As site admin you can install the plugin.
The plugin can be restricted to some projects.

As regular user you can access to the url https://.../artidoc/gpig/
Where `gpig` is the name of a project.

* If you cannot access the project => 404
* If the project does not use Docman/Document => 404
* If the project cannot use the plugin (see plugin restiction) => 404
* In other cases a dummy page is displayed

This allows a contributor not familiar with Tuleap development to have a
working plugin with vue3 application, i18n, vitest, …

Part of story #37542: display a read only Document

Change-Id: I57ad0a780b213c094d7c21ca8b1a4c7458f957fa