feat: Introduce ArtidocDocument
This introduce a new type of document. This type of document is managed
by another plugin than docman/document, so we need to introduce a new
generic type "other" that plugins can use for their will.

To test, you can create an empty document, and in the database change
its `item_type` to `7` and its `item_type` to `artidoc`. The document
should still be displayed in the document tree, and you can search
against it.

For now the icon is the one for trackers, but a tailored icon will be
contributed later.

This is just an introduction of the new document, so it is expected that
not all actions work on it (e.g copy/paste). This will be handled in
dedicated contributions to not clutter this review. It is considered ok
since users have to hack the database to have artidoc documents for now.

Other implementation strategies were envisaged for this:
* transform the `item_type` to a varchar instead of int => issues with
  bisecting if database does not contains the same type in the present
  than in the past.
* use only `item_type` to handle the other type (e.g `7` === 'artidoc',
  `8` === …). The issue is that their is a higher chance of conflicts if
  two different plugins share by mystake the same number for their
  respective types (less chance to have conflicts with 'artidoc').

Part of story #37542: display a read only Document

Change-Id: Id630908dc887c6af4302c9bec5c4555c6032f89a