feat: First display of ArtidocDocument
Go to /artidoc/:id where `:id` is the document with
`(item_type, other_type) = (7, "artidoc")` manually set in the db.

The title of the document is displayed in the page 🥳

If user cannot access to the project, to the document, or the project
does not have nor tracker nor docman services, the user get a not found.

Part of story #37542: display a read only Document

Change-Id: I278b8e4f318744f0f02327a0c8962f5de2611e4d