feat: introduce GET /artidoc/:id/sections
Given the current user can read an artidoc document, then the REST API
will allow to retrieve the content of the document.
`GET /artidoc/:id/sections` where `:id` is the id of the document.

The route will return the list of sections, composed by an artifact
representation. This list is paginated, client will have to do many
calls to retrieve the whole content.

In order to keep the returned result light, the tracker definition is
not returned in each artifact representation. It will be injected in the
application or an additional REST call will have to be done by the
client in order to be able to find the relevant information to compose
the document.

In order to test this patch you will have to hack the database.
In the new table `plugin_artidoc_document` insert entries (list of
artifact id with an arbitrary rank for an artidoc document).

If one of the artifact in the document is not readable by the current
user, then the current call will get a 404.

Part of story #37542: display a read only Document

Change-Id: Ie18a070c10b29a3f613458d3c54d3eebeb73d2f1