feat: Be able to change back to a theme color
part of request #37211 Replace TLP-doc documentation

How to test:
- pnpm --filter=tlp-doc-v2 run storybook
- When you change the theme color to purple and then back to orange, the
orange theme color is applied again (it does not stay purple).
- Note that build-storybook-docs is broken due to presence of `@import`
in the generated CSS for TLP. As those docs are not yet deployed, we can
live with it. Solving this will require re-working the way we build TLP,
especially for the font-awesome fonts part.

Why ?
Switch to Vite builder which is faster than webpack, but also enables us
to import the Theme color variables only once when rendering preview and
then apply them using constructed stylesheet.
Note that the first time I tried Vite when doing the POC for Storybook,
it did not function, which is why we started with webpack in the first
place.

Change-Id: I85c2127a41390fdff0225d75289e32474a18f989