feat: ability to search against artidoc documents
If you have artidoc document in your document manager, then if you start
an advanced search with type = Artidoc, then only them will be displayed
in the interface.

Technical note: since artidoc is a plugin and document is a plugin, the
document app should know nothing about artidoc. This is why the type
AllowedSearchType is removed and we rely now only on what is provided by
the backend during app initialisation.

Part of story #37542: display a read only Document

Change-Id: Ib7fae7da860d9f39e789e9bef2673d3d9b115dca