feat: ignore artidoc documents during project export
If a project contain artidoc documents, then they should not be part of
the XML export (like for wiki documents).

We choose to ignore them for now since we want to validate the main
feature first (viewing/editing a bunch of artifacts as one document).
XML export/import will come later, and we have to think first on how to
deal with artifact linked by artidoc in another projects than the one
being exported (🤯).

No functional changes in the export of folder, wiki (still ignored),
file, embedded, empty, and link.

Part of story #37542: display a read only Document

Change-Id: I1f17d600e8d8387c14b6cde7871e6575fcf22445