feat: display document sections list
A section is a representation of an artifact in an artidoc.

If a artidoc has associated artifacts, then the document view should be display, otherwise it should display the empty state view.
For now, it's a read-only document.

To test go to ${tuleap_url}/artidoc/:id.

Part of story #37542: display a read only Document

Change-Id: Ibd280a1ff967f6bd3a4407dcf55e22afae668dd9