feat: XTS supports @id with = comparison
part of story #36812 search on artifact ids

With this contribution, it is now possible to filter artifacts on their
ids with the = comparison. Please note that, in order to be able to
filter them on this metadata, the tracker needs to have an artifact id
field, and the current user needs to have the read permission on the
field, otherwise, an error message will be displayed.

How to test:
- Using the XTS widget, type 'id' in the editor, then press tab
--> It is autocompleted to `@id`

- type `@id = <artifact_id>` where the artifact you are looking for
  belongs one of the trackers of the current report, then submit.

--> The artifact is found and is displayed (assuming you have the read
permission, and your tracker has an aid field.

Change-Id: Iefe7ac134d39efb4d59025128fc1536a07ab115e