feat: breadcrumbs for artidoc
When we access an artifact document https://example.com/artidoc/:id, we
should have the breadcrumb in the header so that we know where the
document is located (like for an embedded file for example).

If the document is located deep down in the folder hierarchy, then an
ellipsis is displayed to not clutter the interface, and only the last
folders are displayed (like in the vue app).

This is the PHP counterpart of plugins/document/scripts/document/components/Breadcrumb/DocumentBreadcrumb.vue

Part of story #37542: display a read only Document

Change-Id: Idad1550461558fd30e9970d62a2e7a7ada866d8b