fix: request #37568 Do not create FRS data dir in the packaging
It is a source of issues when /var/lib/tuleap/ftp/tuleap is not a
symlink (e.g. when data have been retrieved from a Docker based install)
and you try to upgrade the Tuleap package.

No functional changes for existing instances, however for fresh installs
this drops some parts of FTP shenanigans that are not supported anymore
anyway.

Change-Id: I17d7283cb38213720ac6934e5c8587fe546e38f7