feat: add no access view
The no access view is displayed when the user doesn't have the permision to read at least one of the document artifact.
To test this view, you can setup your database with an artifact not readable by your user, and add it to your artidoc.

Part of story #37542: display a read only Document

Change-Id: I90473689bceaac6395fc9d814ec46bbf770525af