feat: Allow to cut/paste an artidoc
Since frontend of Documents cannot know the existence of artidoc, the
uri to move an item is given by its REST representation. That way we
don't have to guess the uri based on an "unknown" type.

If you have an artidoc document in your Document tree, you can cut/paste
it to move it to another folder. The same for other document types (no
change on that).

Note: the copy/paste deserves its own contribution.

Part of story #37542: display a read only Document

Change-Id: Iff5468fc8a08d454e0d435ef4d703a16f2a14597