fix: wrong order of sections
When we are retrieving artifacts to returns the diferrent sections of an
artidoc, we are using `\Tracker_ArtifactDao::searchByIds()` which don't
have any specific order. This means that there is no guarantee of the
returned collection (it might be based on artifact id or random). Since
we want the section to be ordered we should make sure that the
instantiated artifacts are put in the right order.

Part of story #37542: display a read only Document

Change-Id: If6fbf10c379b28486cdf35934d357611a0d7096b