feat: Copy/paste artidoc document
A user can now copy paste an artidoc document. The content of the
document is the same (same sections, same order).

Note: since we cannot create an artidoc via REST, the REST tests for the
current contribution are quite small. They will be enhanced later.

Part of story #37542: display a read only Document

Change-Id: I7fd0c57d8eb99c84e43b5f2fc98600d0ca5ccb67