chore: remove .tlp-illustration-primary-{fill,stroke}
Those legacy classes are now removed in favor of CSS variables, like
recommended in the documentation.

Note: there is a slight variation of the value of the color between the
legacy and the new one, but it should be ok in most cases.

Part of story #37542: display a read only Document

Change-Id: Id59b22f67d9b389d011b09658f18f4b368259b8b