feat: FTS DB plugin uses UUID instead of auto-increment INT
The main goal is to introduce some extensions of the ForgeUpgrade
BucketDB API to help convert existing tables and to provide a minimal
API to work with UUIDs.

The FTS DB plugin is an adequate playground to start deploying this type
of change. The tables are relatively large but can be entirely reconstitued
from scratch in case of issues.

The forgeupgrade doing the change is written is an agressive way. For tables
with critical data it is going to be preferable to keep the existing PK column
without constraint and defaulting to NULL for new entries.

Part of request #37592: Mitigate risks of data loss in DB

Change-Id: I869f4d6a273c8bb0d25c5500055c1788ce2bae18