feat: introduce first e2e test for artidoc
Now that we can create artidoc documents, and use API to push content
inside them, it is time to introduce e2e test to have a better coverage
of the whole feature.

No functional tests, e2e tests should pass (and can be replayed).

Note: `putFromTuleap` type definition is updated so that we can pass
array instead of just a Record (use `cy.request()` type definition) and
so that we can chain the result.

Part of story #37542: display a read only Document

Change-Id: I97477f203788005732cf111e9ef89864bf2c767b