refactor: remove usage of `any` from cypress-utilities-support
Closes request #37922 Remove usage of `any` from cypress-utilities-support

In the cypress-utilities-support lib, `any` is used to type the content
inside `Response` returned by the different Tuleap API calls wrappers.

Since the `Cypress.Response` and `cy.request` both accept a generic type
to type the returned value, we can drop the `any` type and remove the
eslint-disable comments here and there.

How to test:
- The typecheck should succeed
- Cypress tests should succeed

Change-Id: Ie3f774c70f70cab3392ba06ea25dd7b4b15eef46