feat: Indicates if user can edit document/sections
Frontend should know if user can edit the artidoc document, and if user
can edit a section.

A section can be edited only if user can update the title field AND the description field.

For write permissions on the document, you may need to hack the
permissions table in order to test them.

Part of story #37543: edit Artidoc sections

Change-Id: Id2a3a66786bd8c7931ea0a48d75eb8f92873bed5