refactor: using provide/inject instead of slots and props
For read only document we used slots to manage sections display and skeleton.
But for edit mode, it's not efficient, so now we are using provide/inject to manage sections and loading state like a store.
It's a requirement for the others development on edit mode.

Part of story #37543: edit ArtifactsAsDocument sections

Change-Id: I4f723adf7a3dcfb6a1825a68fc033131a55fa7f0