feat: remove controls in TLP panes stories
We should have controls only for the main story for a component.
Other example stories give an example and are not subject to
customisation.

By moving the args definition to the story instead of the component, there isn't anymore useless/uneeded controls. We just need an excluded control to know which type of story we are dealing with while rendering the component.

Furthermore, the type PaneProps should accept undefined properties (there is no obligation at all to pass all arguments, so we may have a property undefined at runtime, the type should reflect that.

To test, run `pnpm --filter=tlp-doc-v2 run storybook` and go to panes documentation:
* Controls still works for the playground
* Example stories do not have any controls

Part of request #37211: Replace TLP-doc documentation

Change-Id: Ifa5c5759849b6e1ca0700b8f2f8d8f9894e1a4d7