feat: Burnup cache checker uses count cache if needed
Part of request #37921 Burnup/Burndown cache generation fail

Burnup can be generated from effort or count of subelement. But when we
check if cache is valid, we only check effort cache (which is generated
even if not needed)

*Testing:*

You will need to hack the database

Have a burnup based on count of subelement. Generate its cache. It
should have filled table
plugin_agiledashboard_tracker_field_burnup_cache_subelements.

Remove one (or more) entry of the table: before the patch the cache is
still considered valid. With the patch the cache is considered invalid
and generated again

Change-Id: Ibdc4ab30e98e5d9104b267a4e3eaff3f9e6740e1