feat: Feature flag for artidoc edition
Feature will not be ready for the current release, we don't want early
birds to see a broken feature.

In order to be able to edit a document, a user must have write
permissions on the document AND the feature flag must be activated.

To set the feature flag in your dev environment:
tuleap config-set feature_flag_enable_artidoc_edition 1

Note: since the edit mode is not yet contributed, you will have to see
at the source of the html page of an artidoc document to assert the
correctness of this contribution: the data attribute
`data-can-user-edit-document` should reflect the value of the feature
flag.

Part of story #37543: edit Artidoc sections

Change-Id: Ibfb3526325f1f1c5f94ebb5a2fb4a97ef61e529f