feat: introduce section id
We will need a section id later (for example to refresh the content of
a section after the corresponding artifact update, so that we can
retrieve the post processed value of the description).

We cannot use the artifact id as section id since we don't want to open
the door to breaking changes the day we have other thing than artifact
as section (eg free text).

The introduced id is a UUID, as recommended by
[ADR-0028](adr/0028-prevent-data-loss.md).

No functional changes, the id is visible in the REST representation of
sections.

Part of story #37543: edit Artidoc sections

Change-Id: I2eef93c8c7194bbb9c8da801f91e632ab70e16e9