feat: Display CKEditor in the user locale
The interface of the editor should match user's preferences.

Implementation note: it appears that *all* languages supported by ckeditor are being build as assets. This generates a lot of unwanted files. It would be great to limit the assets to the one supported by Tuleap in local.inc.dist (at build time). For now we let it as is, considering that having unused built files is better than having to manually list the supported languages in the artidoc application.

Part of story #37543: edit Artidoc sections

Co-authored-by: Thomas Gorka <thomas.gorka@enalean.com>
Co-authored-by: Nicolas Terray <nicolas.terray@enalean.com>

Change-Id: Id55d1225c7c6fed41f6e1f69ff39f6ef1f467427