feat: GET artidoc_sections/:id
We can now retrieve the content of one section only (until now we had to
query the whole document to know the content of a section).

This will be usefull to refresh the new content of a section after its
edition (Needed to get the post processed value, for cross references
links in the new content).

The :id parameter of the route is the id (uuid) of the section that can
be retrieved in the array of sections of the document.

Part of story #37543: edit Artidoc sections

Change-Id: I453df025c9b05d51e3d77804dfe4ba456cce02cc