fix: request #37954 WASM User error is malformed
As the key is wrong, the Valinor mapper failed to map and we get an
execution exception with the mapping error instead of the real error.

Fix on the rust side to be consistent with the internal error struct.

*Testing:*

With a tracker consisting of only a text field, upload the TF add
comment. When creating a new artifact, the TF should fail and if you
checks the logs there is a new line in error state and the error message
is :

    An error occurred while running artifact post action WASM function: Exited with i32 exit status 1
    |                Tuleap Function error                             |         WASM error

Change-Id: I5e372b165373a289dabcee7ebb5f9033c82e05cd