fix: Cannot merge PR via the web UI
Git 2.44.1 has added restrictions when cloning local repositories.
In our situation it is fine as we only handle repositories under Tuleap
control (and Tuleap builds all commands) so we consider everything to be
a safe directory. In a ideal world the merges would be done in the same
context than the repository and in memory to avoid the need of cloning
the repo elsewhere.

Part of request #37952 Git: 2.44.0 -> 2.44.1

Change-Id: Ic8986f17b759d816f24dc2c9e47f8e7ccef7d710