feat: Be able to select predefined time periods
Part of story #37941 Timetracking management widget

Added a predefined period selector to facilitate date selection.
We use the library of predefined periods, previously extracted in timetacking, for consistency.
The time period selected is in the props, because we're going to save it in the database to enable persistence on it.
No data is currently retrieved, because this commit only allows you to select and display dates through the predefined period selector.

The widget is currently disabled by a feature flag. To set the feature flag to 1:
tuleap config-set feature_flag_allow_timetracking_management_widget 1

How to test:
Click on the dates. You can select predefined time periods in the selector.
The dates in the date selectors should change automatically according to the predefined period.
When you search and click again on the dates, then the last predefined period selection should be retained.

Change-Id: I8dda1a35af17a9b826f53279e3f8015508763403