feat: Highlight section depending on its status
When the section is in read-only, its backgrond should be white.
When the section is beeing edited, its background should be blueish.
When the section is beeing saved, its background should blink
white/blueish to indicate to the user that something is happening.[0]
When the section is finally saved, its background should be greenish for
a short amount of time before returning back to white.[0]
When the section cannot be saved because of error, its background should
be redish.[1]

[0] Same behavior than kanban card.
[1] No error message is displayed yet, it will be part of a dedicated
contribution.

Note: Took the opportunity to rename getIsEditMode to isXxxx since it
returns a boolean.

Part of story #37543: edit Artidoc sections

Change-Id: I9635c49c3bb1bc7c3f30a589eaba63f7f6504adf