feat: Edition of title
When a section is edited, user can now edit its title alongside its
description.

This works if the semantic is bound to a string field or a text field.
If text field, then text, html, commonmark will be forced to text format
on save.

Part of story #37543: edit Artidoc sections

Change-Id: Ibdfaaac52243c48498119e52899656427e606c7f