feat: Introduce Query and Selectable
Part of story #38263 Choose my own columns based on field name (numeric, text, dates)

Class Query is a wrapper for TQL Query which contains both the condition
and the list of selected fields.

As before we only have the condition, this class was not necessary, but
with the addition of the SELECT we need it (and also in the future with
FROM and ORDER BY)

*Testing:*

You can write tql queries as before.

- If you try to use select it should returns a syntax error
- Set the feature flag `feature_flag_enable_tql_select` to 1, you should
  be able to use the select syntax

Change-Id: Ib97bac9b827a914bbf18d247b60f237b546789c7