feat: Concurrent edition of sections in artidoc
Given Alice starts to edit a section,
And Bob concurrently save a new version of the corresponding artifact,
When Alice clicks on save,
Then a warning is displayed asking to either force save or load the new
section.

When Alice clicks on force save,
Then there is no check for concurrent update.

When Alice clicks on load new section,
Then the editor is closed and the latest version of the section is
displayed,
And if Alice edits again the section,
Then the editor displays the latest version of the section to edit.

Part of story #37543: edit Artidoc sections

Change-Id: Ibea291f444b88adc6c8659a88317d50a1375c6f3