refactor: Convert reading mode to composition API
part of story #38263 Choose my own columns based on field name
(numeric,text, dates)

How to test:
No functional change expected in cross-tracker search widgets. The
"reading mode" is when a report (selected trackers + query) is already
saved, or modified but not saved yet. It shows a read-only list of
trackers and a read-only query and a table listing artifacts that match
the query.

There is no longer an underline before/after the text in the <a> that
links to the project URI, the title link or the users links.

Why?
Class components are no longer supported in Vue 3. We must switch to
composition API to upgrade to Vue 3.

Note: Type Wrapper<Vue, Element> is used because in this middle ground
between Vue 2 and 3, we cannot use the real type yet.

Change-Id: I75d46b05172dbe0aafc269cd325a3836f01e3a3f