chore: Add base structure to build TQL select SQL
Part of story #38263 Choose my own columns based on field name (numeric, text, dates)

To be able to build easily the SQL query for the select part of TQL
query, we need a base class structure. This structure is pretty similar
to the one for the condition.

In this commit this structure does nothing and returns a query which
simply returns the artifacts ids.

On the user side nothing change

*No functional changes expected*

Change-Id: Iad1eda4d4a0536f83ff5f495f52fc42fe37420a3