feat: Dedicated message in case of 404 while editing a section
When we try to save a section, the update of the artifact might fail in
a 404 (the artifact has been deleted meanwhile, or the project has been
deleted, …). In that case we should incitate the user to refresh the
current page since the document currently displayed is not in a clean
state.

Part of story #37543: edit Artidoc sections

Change-Id: I86453c33a1efc071055909df182237fa102d24b0