fix: Unselecting trackers should be accessible by keyboard
part of story #38263 Choose my own columns based on field name
(numeric,text, dates)

How to test:
- In cross-tracker search widget, as project admin, after clicking on
the widget, you can select trackers. The badge that shows a selected
tracker now has a button (instead of just an icon) to remove said
tracker from the selection. This button should be accessible by
navigation with TAB. It should be possible to unselect trackers using
the keyboard alone: press enter or space when the unselect button is
focused.
- Same scenario for timetracking overview widget.
- No functional change expected for lazybox with multiple values.

Change-Id: I2e18bedfc385209dbd72ec7f6eaacee6df638850