feat: handle 403 while editing a section
When PUT /artifacts/:id, it means that user don't have access to the
artifact resource (e.g., cannot view the artifact, or cannot access the
tracker, or cannot access the project). Therefore if a user tries to
update a section and get a 403, we should reload the page to remove the
section from the display, like for NotFound error.

Not being able to update a field ends up in a 400, so this will be part
of a dedicated contribution.

Part of story #37543: edit Artidoc sections

Change-Id: Ief5c2ec41d90b08a0aa875e56f4a267c0f70e5a6