chore: Close request #38274 Missing translation in Git commit details
How to test:
Browse to the details page of a given commit in a Git repository, the
list of changed files is displayed; next to each file, there is a "View
file" button. Its label should be translated in French.

Also, the icon for the "List" button that shows the list of changed
files (the current page essentially) should now have correct margin (and
not be glued to the label).

Notes:
Also changed all icons to add aria-hidden attribute and Font Awesome 6
classes. All icons should still display.

Change-Id: Idae413757150b0cee8f2df79055cd08079851207