feat: handle 400 and 500 errors
There are different cause to get a 400 or 500 while trying to update an
artifact:
* The administrator just updated the tracker structure and permissions
  do not allow anymore to update the title or description
* Fields dependencies have just been created and the existing artifact
  does not respect them
* Internal server error
* …

We cannot differentiate one to each other so we have to display a
generic message. The user can still access to its unsaved modifications
and can open the artifact in a new tab, so that they can potentially fix
the situation.

Bonus: if the user save the section without updating anything, then the
section does not blink forever.

Part of story #37543: edit Artidoc sections

Change-Id: I114a0ac25f1dcf83ae57d24212de3de5aad3d2a5