feat: Match label with input
part of request #37211 Replace TLP-doc documentation

How to test:
pnpm --filter=tlp-doc-v2 run storybook

Why?
When I click on the label above an input, the input should be focused.
To achieve that, we should set an "id" attribute on our inputs (and
textarea) and set a "for" attribute on our labels. "id" and "for" values
must be the same, and must be unique on a given page.

Change-Id: Ibe33818b154723fbd23ec2a837079bfa51b5804e