feat: Edit/save buttons should be sticky
When the user edits a section, they should not have to scroll to reach
the edit, save, or cancel button.

Save and cancel buttons are now at the bottom of the section (the same
than what is done on the artifact view).

If on save there is an error (NotFound, GenericError, or Outdated, then
the message is displayed in the footer, on top of the buttons.

Part of story #37543: edit Artidoc sections

Change-Id: Ic4ae7e0664ace33f6985feedd6c7c143e0c1bb59