chore: TS Types and Interfaces should always be PascalCased
Implements request #38277 Enforce naming TS types using PascalCase

In the codebase, our TypeScript types and interfaces are named using the PascalCase.
It is an implicit convention we are used to in the Tuleap team, so let's enforce it with an eslint rule.
FTR, some types using the snake_case have been introduced in the codebase, let's fix them in the meantime.

How to test:
- Pick up a TS file
- Define a `type whatever = {}` and an interface `interface something {}` in this file.
- Commit

--> Eslint complains about the type of case used to name them.

Change-Id: Ia7543c12f01d553c7a40183399a72588a3f68e9d