fix: request #38284 Do not allow end users to update project references created/managed by plugins
It is no longer possible to edit tracker references to avoid some
issues.

Quick reproduction scenario of a situation where this is problematic:
1. In project A, create a tracker
2. Edit the reference that has been created automatically (just click update,
   you do not even need to change something)
3. Try to create a new project from this project, the project creation will fail
   because the name is no longer associated with the tracker plugin (and so we
   cannot create the tracker since the name is already reserved)

Note this is far from perfect and the whole reference system needs some
work (at least on the administration parts) but I'm a bit afraid to more
invasive changes right now, a lot of things in this place is still tied
to the project ID 100 or to site admin rights and changing things might
break (business critical) features. Removing the possibility of edition
for these services at least remove what seems to be the most obvious
undocumented footgun.

Change-Id: Ie96ebbc17ff90139a3a8f5e29f7db5eaa2672b0b